// frontend/flume/src/types.ts
export type ControlData = { [controlName: string]: unknown };
export type InputData = { [portName: string]: ControlData };

export type ControlTypes =
    | "text"
    | "number"
    | "select"
    | "checkbox"
    | "multiselect"
    | "custom";

export type ValueSetter = (newData: unknown, oldData: unknown) => unknown;

export interface GenericControl {
    type: ControlTypes;
    label: string;
    name: string;
    defaultValue: unknown;
    setValue?: ValueSetter;
}

export interface TextControl extends GenericControl {
    type: "text";
    defaultValue: string;
}

export interface SelectOption {
    label: string;
    value: string;
    description?: string;
    sortIndex?: number;
    node?: NodeType;
    internalType?: "comment";
}
// ...
export type ControlRenderCallback = (
    data: unknown,
    onChange: (newData: unknown) => void,
    context: unknown,
    redraw: () => void,
    portProps: {
        label: string;
        name: string;
        portName: string;
        inputLabel: string;
        defaultValue: unknown;
    },
    controlData: ControlData
) => ReactNode;
// ...
export interface PortType {
    /**
     * A unique string identifier for the port. Preferred to be camelCased.
     */
    type: string;
    /**
     * A default string identifier used when the port is constructed.
     */
    name: string;
    /**
     * A default human-readable label for the port.
     */
    label: string;
    /**
     * When true the port will not render its controls.
     *
     * @defaultValue false
     */
    noControls: boolean;
    /**
     * The color of the port. Should be one of the colors defined by the Colors type.
     */
    color: Colors;
    /**
     * If true the ports controls will render but the actual port will not. This disallows connections.
     *
     * @defaultValue false
     */
    hidePort: boolean;
    /**
     * An array of controls to render on the port.
     */
    controls: Control[];
    /**
     * An array of port type strings. Only port types included in this array will be allowed to connect to this port. By default ports always accept their own type.
     */
    acceptTypes?: string[];
}

export type PortTypeMap = { [portType: string]: PortType };

export type PortTypeBuilder = (config?: Partial<PortType>) => PortType;

export interface PortTypeConfig extends Partial<PortType> {
    type: string;
    name: string;
}

export type TransputType = "input" | "output";

export type TransputBuilder = (
    inputData: InputData,
    connections: Connections,
    context: unknown
) => PortType[];

export interface NodeType {
    /**
     * A unique randomly-generated string identifier for the node.
     */
    id: string;
    /**
     * A unique string identifier for the node. Preferred to be camelCased.
     */
    type: string;
    /**
     * A human-readable label for the node.
     */
    label: string;
    /**
     * A human-readable description for the node. Renders in the "Add Node" context menu.
     * Optional.
     */
    description?: string;
    /**
     * If false the node may not be added to the canvas.
     * Optional; when omitted, nodes are addable.
     *
     * @defaultValue true
     */
    addable?: boolean;
    /**
     * If false the node may not be removed from the canvas.
     * Optional; when omitted, nodes are deletable.
     *
     * @defaultValue true
     */
    deletable?: boolean;
    inputs: PortType[] | TransputBuilder;
    outputs: PortType[] | TransputBuilder;
    initialWidth?: number;
    sortIndex?: number;
    root?: boolean;
}